Added Name field for Themes, modified speed to a double for WorkOut

+++ database/migrations/2026_05_29_113509_create_work_outs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id');
             $table->integer('length');
-            $table->integer('speed');
+            $table->double('speed');
             $table->string('type');
             $table->integer('points');
             $table->timestamps();
